Use Berkeley DB instead of SQLite to store the data blocks
From the updated README:

The file system initially stored everything in a single SQLite database, but it
turned out that after the database grew beyond 8 GB the write speed would drop
from 8-12 MB/s to 2-3 MB/s. Therefor the file system now uses a second database
to store the data blocks. Berkeley DB is used for this database because it's
meant to be used as a key/value store and doesn't require escaping binary data.