Improved flexibility of SQLite database initialization
Previously if the SQLite database file didn't exist yet it
would be initialized for first use, otherwise the database
was presumed to already be properly initialized. This
approach made it impossible to use DedupFS with for example
Python's tempfile.mkstemp() function.

I've now changed the file system's SQL initialization code
to always run and I've added IF NOT EXISTS and OR IGNORE
clauses so that the initialization code is ignored after
the first run. This shouldn't take more than a second or so
on each initialization but it makes the code more robust.